+++ server/routes/auth.js
@@ -178,37 +178,6 @@ router.post('/register/company', function (req, res, next) {
 
 });
 
-// update user username and password by id
-router.put('/update/user/:id', function (req, res, next) {
-    if (req.body.username && req.body.password && req.body.id) {
-        bcrypt.hash(req.body.password, 10, (err, hash) => {
-            users.updateUser(req.body, hash, (dberr, dbRes) => {
-                if (dberr) {
-                    errors.errorCode(dberr, res)
-                } else {
-                    users.getUserData(req.body.username, (dberr, dbResult) => {
-                        if (dberr) {
-                            // If error return error
-                            errors.errorCode(dberr, res)
-                        } else {
-                            // Return company details and json web token
-                            // Get json web token
-                            const token = generateAccessToken({ userData: dbResult[0] });
-                            dbResult[0].token = token
-                            // Delete password field
-                            delete dbResult[0].password
-                            // Return
-                            res.send(dbResult[0])
-                        }
-                    })
-                }
-            })
-        })
-    } else {
-        errors.errorCode("missing params", res)
-    }
-});
-
